ai-engine/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors
pillow
opencv-python
pydicom
numpy
scikit-image
python-multipart
python-dotenv
bitsandbytes
